backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai-agents
python-multipart
python-dotenv
